rc: Do not flush pages on every rc signal, as this could impact general performance (experimental change)
The purge currently occurs every time that the root init process gets signalled.  Typically, this would
be when a service gets restarted through rc_notify, so for normal router usage, this would mostly occur
when applying a change on the webui.  Therefore, I am unsure at this point if there is any real benefit
to be gained, versus the chance that this purge might have been initially added by BCM to deal with cache
coeherence issues that are specific to either 2.6.36 or SMP.  And on the other hand, the Linux documentation
clearly states that this proc control should ideally only be used for debugging/testing purposes.

So, testing this patch in for a while, will see if it has any real (postiive or negative) impact.